Allow Custom Elements API element extensions
The Custom Elements API supports extending existing elements, including both custom elements and built-in elements.  In order for this to work, however, the custom element class must extend the actual class of the element being extended.  This commit allows passing the name of the element being extended and its class.